\name{M1}
\docType{data}
\alias{M1}
\title{Matrix of simulated ordinal data}
\description{
 This is a toy dataset for running simple examples.
}
\usage{M1}

\format{ A mixed type data matrix with 50 lines and 120 columns. There are 40 categorical variables, 40 continuous variables, and 40 ordinal variables.
}

\keyword{datsaets}